% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRdrawDRC.R
\name{GRdrawDRC}
\alias{GRdrawDRC}
\title{Dose-Response Curves}
\usage{
GRdrawDRC(fitData, experiments = "all", min = "auto", max = "auto",
  points = TRUE, curves = TRUE, plotly = TRUE)
}
\arguments{
\item{fitData}{an element of class GRfit, generated by the GRfit function.}

\item{experiments}{the names of the experiments to plot (or "all")}

\item{min}{the minimum concentration to plot (for curves)}

\item{max}{the maximum concentration to plot (for curves)}

\item{points}{a logical value indicating whether points (individual GR
values) will be plotted}

\item{curves}{a logical value indicating whether sigmoidal dose-response
curves will be plotted}

\item{plotly}{a logical value indicating whether to output a ggplot2 graph
or a ggplotly graph}
}
\value{
ggplot2 or ggplotly graphs of Growth-rate inhibition dose-response
curves
}
\description{
This function plots the Growth-rate inhibition dose-response curves for a
given set of data.
}
\details{
Given a SummarizedExperiment object created \code{\link{GRfit}},
this function plots these GR values (versus concentration) and/or the
sigmoidal curves fitted to the sets of points. The results can be viewed
in a static ggplot image or an interactive plotly graph.

The "min" and "max" parameters control the concentration values for which
the curves are plotted. They are automatically set to the minimum and
maximum concentrations of the data, but can be set by the user as well.
"min" and "max" take raw values (not log transformed) for concentration.

By default, curves and points are plotted for all experiments. To specify
a smaller set of experiments, use the "experiments" parameter. To see the
names of individual experiments, use
\code{GRfitObject$parameter_table$experiment}. See the examples below.
}
\examples{
# Load Case A (example 1) input
data("inputCaseA")
# Run GRfit function with case = "A"
drc_output = GRfit(inputCaseA,
groupingVariables = c('cell_line','agent'))
GRdrawDRC(drc_output, experiments = c('BT20 drugA', 'MCF10A drugA',
'MCF7 drugA'), min = 10^(-4), max = 10^2)
GRdrawDRC(drc_output, plotly = FALSE)
}
\author{
Nicholas Clark
}
\seealso{
To create the object needed for this function, see
\code{\link{GRfit}}. For other visualizations, see \code{\link{GRbox}} and
\code{\link{GRscatter}}. For online GR calculator and browser, see
\url{http://www.grcalculator.org}.
}

