% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRfunctions.R
\name{GRfit}
\alias{GRfit}
\title{Extract GR parameters from a dataset}
\usage{
GRfit(inputData, groupingVariables, GRtable = "both", force = F, cap = F,
  case = "A", concentration = "concentration", cell_count = "cell_count",
  time = "time", cell_count__time0 = "cell_count__time0",
  cell_count__ctrl = "cell_count__ctrl")
}
\arguments{
\item{inputData}{a data table in one of the specified formats (Case A or
Case C). See details below or
\url{https://github.com/sorgerlab/gr50_tools/blob/master/README.md}}

\item{groupingVariables}{a vector of column names from inputData. All of the
columns in inputData except for those identified here will be averaged over.}

\item{GRtable}{a character string ("param", "GR", or "both") identifying
whether to return, respectively, only the table of parameters, only the
table of GR values, or both tables. Default is "both".}

\item{force}{a logical value indicating whether to attempt to "force" a
sigmoidal fit, i.e. whether to allow fits with F-test p-values greater
than .05}

\item{cap}{a logical value indicating whether to cap GR values in between
-1 and 1 (the range of the GR dose-response curve). If true, all GR values
greater than 1 will be set to 1 and all values less than -1 will be set
to -1.}

\item{case}{either "A" or "C", indicating the format of the input data. See
here \url{https://github.com/sorgerlab/gr50_tools/blob/master/README.md}
or details below for descriptions of these formats.}

\item{concentration}{(Case A and Case C) the name of the column with
concentration values (not log transformed) of the perturbagen on which
dose-response curves will be evaluated}

\item{cell_count}{(Case A and Case C) the name of the column with the
measure of cell number (or a surrogate of cell number such as CellTiter-Glo®
staining) after treatment}

\item{time}{(Case C) the name of the column with the time at which a cell
count is observed}

\item{cell_count__time0}{(Case A) the name of the column with Time 0 cell
counts - the measure of cell number in untreated wells grown in parallel
until the time of treatment}

\item{cell_count__ctrl}{(Case A) the name of the column with the Control
cell count - the measure of cell number in control (e.g. untreated or
DMSO-treated) wells from the same plate}
}
\value{
By default, a list of three elements is returned: 1) a data table of
the original data, converted to the style of Case A, with GR values for each
experiment 2) a data table of GR metrics parameters (GR50, GRmax, etc.) as
well as goodness of fit measures. 3) the vector of grouping variables (this
is used for other functions such as \code{\link{GRdrawDRC}}). If GRtable is
equal to "GR", then only (1) and (3) are returned. If GRtable is set to
"param", then only (2) and (3) are returned.
}
\description{
This function takes in a dataset with information about concentration,
cell counts over time, and additional grouping variables for a dose-response assay
and calculates growth-rate inhibition (GR) metrics for each experiment in
the dataset. The data must be in a specific format: either that specified
by case "A" or case "C" described here
\url{https://github.com/sorgerlab/gr50_tools/blob/master/README.md} and in
the details below, although the column names of the key variables
(concentration, cell_count, etc.) can be specified in the command.
}
\details{
Calculation of GR values is performed by the function \code{.GRcalculate}
according to the "Online Methods" section of Hafner and Niepel et al.
(\url{http://dx.doi.org/10.1038/nmeth.3853}).

The fitting of the logistic curve is performed by the \code{.GRlogisticFit}
function, which calls the \code{drm} function from the \code{drc} package
to solve for the curve parameters. The GR curve fit function is
given by f(c) = GRinf + (1 - GRinf)/(1 + (c/GEC50)^Hill) where c is
concentration. The fit is performed under following constraints: Hill slope
in [.1, 5], GRinf in [-1, 1], and GEC50 in [min(c)*1e-2, max(c)*1e2] (c is
concentration). The initial conditions for the fitting algorithm are Hill
slope = 2, GRinf = 0.1 and GEC50 = median(c).

The parameters of the curve for each experiment are fitted separately. An
F-test is used to compare the sigmoidal fit to a flat line fit. If the
p-value of the F-test is less than .05, the sigmoidal fit is accepted. If
the p-value is greater than or equal to .05, a flat horizontal line fit is
given, with y equal to the mean of the GR values. In the parameters data
table, for each flat fit, GEC50 is set to 0, Hill is set to 0.01, GRinf is
set to the y value of the flat fit (the mean of the GR values), and GR50 is
set to +/-Inf depending on whether GRinf is greater or less than .5.
}
\note{
To see the underlying code, use (\code{getAnywhere(.GRlogistic_3u)}),
(\code{getAnywhere(.GRcalculate)}), and (\code{getAnywhere(.GRlogisticFit)})
}
\examples{
# Load Case A (example 1) input
data("inputCaseA")
head(inputCaseA)
# Run GRfit function with case = "A"
output1 = GRfit(inputData = inputCaseA, groupingVariables = c('cell_line','agent', 'perturbation','replicate', 'time'))
# See parameter table
head(output1$parameter_table)
# See GR values table
head(output1$gr_table)
# Load Case C (example 4) input
# Same data, different format
data("inputCaseC")
head(inputCaseC)
output4 = GRfit(inputData = inputCaseC, groupingVariables = c('cell_line','agent', 'perturbation','replicate', 'time'), case = "C")

}
\author{
Nicholas Clark
}
\references{
Hafner, M., Niepel, M. Chung, M. and Sorger, P.K., "Growth Rate Inhibition Metrics Correct For Confounders In Measuring Sensitivity To Cancer Drugs". \emph{Nature Methods} 13.6 (2016): 521-527.
\link{http://dx.doi.org/10.1038/nmeth.3853}
}
\seealso{
See \code{\link{drm}} for the general logistic fit function that
solves for the parameters GRinf, GEC50, and Hill slope. See \code{\link{drmc}} for
options of this function. Use the functions \code{\link{GRdrawDRC}},
\code{\link{GRbox}}, and \code{\link{GRscatter}} to create visualizations
using the output from this function. For online GR calculator and browser, see
\url{http://www.grcalculator.org}.
}

