% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputCaseA}
\alias{inputCaseA}
\title{Simulated dose-response assay data (Case A)}
\format{A data frame with 1008 rows and 9 variables:
\itemize{
  \item cell_line: the cell-line used (MCF10A, MCF7, BT20)
  \item agent: the drug used (drugA, drugB, drugC, drugD)
  \item pertubation: an example of another key variable (e.g different
  media, a co-treatment, etc.) (0, 1)
  \item replicate: replicate number, (1, 2, 3)
  \item time: time of assay measured in hours (48, 72)
  \item concentration: concentration of the perturbagen on which dose-response curves
   will be evaluated (not log transformed)
  \item cell_count: the measure of cell number (or a surrogate of cell number) after treatment at the end of the assay
  \item cell_count__ctrl: the measure of cell number in control
  (e.g. untreated or DMSO-treated) wells from the same plate at the end of the assay
  \item cell_count__time0: the measure of cell number in untreated wells grown in
   parallel until the time of treatment
}}
\source{
\url{https://github.com/sorgerlab/gr50_tools/blob/master/INPUT/toy_example_input1.tsv}
}
\usage{
inputCaseA
}
\value{
An example dataset in the form of "Case A" generated for use with \code{\link{GRfit}}
}
\description{
This is the dataset corresponding to Case A (example 1 in MATLAB scripts).
See this page for details of the different cases of input considered:
\url{https://github.com/sorgerlab/gr50_tools/blob/master/README.md}
The data was generated by the python script found here:
\url{https://github.com/sorgerlab/gr50_tools/blob/master/examples/generate_data/generate_data.py}
}
\keyword{datasets}

