% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputCaseC}
\alias{inputCaseC}
\title{Simulated dose-response assay data (Case C)}
\format{A data frame with 1352 rows and 7 variables:
\itemize{
  \item cell_line: the cell-line used (MCF10A, MCF7, BT20)
  \item agent: the drug used (-, drugA, drugB, drugC, drugD)
  \item pertubation: an example of another key variable (e.g different
  media, a co-treatment, etc.) (0, 1)
  \item replicate: replicate number, (1, 2, 3)
  \item time: time of assay measured in hours (0, 48, 72)
  \item concentration: concentration of the perturbagen on which dose-response curves
   will be evaluated (not log transformed)
  \item cell_count: measure of cell number or a surrogate of the number of cells.
}}
\source{
\url{https://github.com/sorgerlab/gr50_tools/blob/master/INPUT/toy_example_input4.tsv}
}
\usage{
inputCaseC
}
\description{
This is the dataset corresponding to Case C (example 4 in MATLAB scripts).
The dataset is equivalent to that of Case A, but in a different form.
See this page for details of the different cases of input considered:
\url{https://github.com/sorgerlab/gr50_tools/blob/master/README.md}
The data was generated by the python script found here:
\url{https://github.com/sorgerlab/gr50_tools/blob/master/examples/generate_data/generate_data.py}
}
\keyword{datasets}

