% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRbox.R
\name{GRbox}
\alias{GRbox}
\title{Boxplots of a given GR metric}
\usage{
GRbox(fitData, GRmetric, groupVariable, pointColor, factors = "all",
  plotly = TRUE)
}
\arguments{
\item{fitData}{an element of class GRfit, generated by the GRfit function.}

\item{GRmetric}{the GR metric that want to use for the boxplot (GR50, GRinf,
Hill, GRmax, GEC50, or GR_AOC)}

\item{groupVariable}{the name of the variable from data (e.g. drug,
cell-line, etc.) to select factors from.}

\item{pointColor}{a variable that defines the coloring of the points
overlayed on the boxplot.}

\item{factors}{a vector of values of "groupVariable" of data that define
which variables to make boxplots for. By default, a separate boxplot is made
for each unique value of groupVariable.}

\item{plotly}{a logical value indicating whether to output a ggplot2 graph
or an interactive ggplotly graph}
}
\value{
ggplot2 or ggplotly boxplots of the factors along the x-axis, with
points colored by the given variable.
}
\description{
This function takes in a data frame generated by GRfit and creates boxplots
according to the parameters below.
}
\details{
Given a SummarizedExperiment object created by \code{\link{GRfit}},
this function creates boxplots of a given GR metric (GR50, GRmax, etc.)
for values of the grouping variable. The results can be viewed in a static
ggplot image or an interactive plotly graph.

By default, a boxplot is created for all unique
values of the grouping variable. The "factors" parameter can be used to
specify a smaller subset of values for which to create boxplots.
Points are overlayed on the boxplots and
they can be colored by the variable specified in the pointColor parameter.
If pointColor is set to NULL, the points will all be black. The results can
be viewed in a static ggplot image or an interactive plotly graph.
}
\examples{
# Load Case A (example 1) input
data("inputCaseA")
head(inputCaseA)
# Run GRfit function with case = "A"
output1 = GRfit(inputData = inputCaseA,
groupingVariables = c('cell_line','agent', 'perturbation','replicate',
'time'))
GRbox(output1, GRmetric ='GRinf',
groupVariable = 'cell_line', pointColor = 'agent' , factors = c('BT20',
'MCF10A'))
GRbox(output1, GRmetric ='GRinf',
groupVariable = 'cell_line', pointColor = 'cell_line' ,
factors = c('BT20', 'MCF10A'), plotly = FALSE)
}
\author{
Nicholas Clark
}
\seealso{
To create the object needed for this function, see
\code{\link{GRfit}}. For other visualizations, see \code{\link{GRdrawDRC}}
and \code{\link{GRscatter}}. For online GR calculator and browser, see
\url{http://www.grcalculator.org}.
}

